# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Politics Project",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/datarecords/{datarecord_id}": {
      "get": {
        "tags": [
          "datarecords",
          "datarecords"
        ],
        "summary": "Get Datarecord",
        "description": "Get a specific DataRecord.",
        "operationId": "get_datarecord",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          },
          {
            "name": "datarecord_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Datarecord Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/datarecords/by_datasource/{datasource_id}": {
      "get": {
        "tags": [
          "datarecords",
          "datarecords"
        ],
        "summary": "List Datarecords",
        "description": "List DataRecords for a specific DataSource.",
        "operationId": "list_datarecords",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          },
          {
            "name": "datasource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Datasource Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataRecordRead"
                  },
                  "title": "Response Datarecords-List Datarecords"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/datarecords/by_datasource/{datasource_id}/records": {
      "post": {
        "tags": [
          "datarecords",
          "datarecords"
        ],
        "summary": "Append Record",
        "description": "Append a record to a DataSource.",
        "operationId": "append_record",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workspace Id"
            }
          },
          {
            "name": "datasource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Datasource Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendRecordInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppendRecordInput": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "The text content or URL to append"
          },
          "content_type": {
            "type": "string",
            "enum": [
              "text",
              "url"
            ],
            "title": "Content Type",
            "description": "Type of content being appended"
          },
          "event_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Timestamp",
            "description": "Optional ISO 8601 timestamp for the event"
          }
        },
        "type": "object",
        "required": [
          "content",
          "content_type"
        ],
        "title": "AppendRecordInput",
        "description": "Input model for appending a record to a datasource."
      },
...
      "DataRecordRead": {
        "properties": {
          "text_content": {
            "type": "string",
            "title": "Text Content"
          },
          "source_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Metadata"
          },
          "event_timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Timestamp"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "datasource_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datasource Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          }
        },
        "type": "object",
        "required": [
          "text_content",
          "id",
          "created_at"
        ],
        "title": "DataRecordRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}